# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MailMAster API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/subscribers": {
            "get": {
                "tags": [
                    "Subscribers"
                ],
                "summary": "Get all subscribers",
                "operationId": "7a4ec4a0bed9cb1bb0f252aa5fde1ac4",
                "responses": {
                    "200": {
                        "description": "A list of subscribers"
                    }
                }
            },
            "post": {
                "tags": [
                    "Subscribers"
                ],
                "summary": "Create a new subscriber",
                "operationId": "cdbdfb57ab947416e0ecf36c82ed29e8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "newsletter_id"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "newsletter_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Subscriber created successfully"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/subscribers/{id}": {
            "get": {
                "tags": [
                    "Subscribers"
                ],
                "summary": "Get a single subscriber",
                "operationId": "cb2ebc56c624405a54492302dbb13aee",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Details of a Subscriber"
                    },
                    "404": {
                        "description": "Subscriber not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Subscribers"
                ],
                "summary": "Update a subscriber",
                "operationId": "f7edc93245402315d3f961da3ec023a6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "newsletter_id"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "newsletter_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Subscriber updated successfully"
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "404": {
                        "description": "Subscriber not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Subscribers"
                ],
                "summary": "Delete a subscriber",
                "operationId": "d5182899aaacc15b1b2a04f7808dd746",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subscriber deleted successfully"
                    },
                    "404": {
                        "description": "Subscriber not found"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}